// src/app/core/api/generated/schemas/topProductDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts API
 * OpenAPI spec version: 1.0
 */
import type { TopProductDtoCategory } from "./topProductDtoCategory";

export interface TopProductDto {
  product_id: number;
  title: string;
  sku: string;
  /** @nullable */
  category?: TopProductDtoCategory;
  sold: number;
  revenue: number;
}